requirements.txt
fastapi
uvicorn
pandas
networkx
python-multipart
python-dotenv
torch
torch-geometric
scikit-learn
joblib
